Extra requirements-file, new route, cleaner routes * Added requirements-full.txt which holds requirements for not just the project but also indirect requirements like virtual environments and (eventually) Docker-files. * Cleaned up all routes.     * Centralized prefixes in main.py.     * Moved helper code to new auth.py-file in the new dependencies-folder.     * Imported the helper codes back to the post-routes. * Added route to get all posts from a user, which will be useful on things like profile and user-post overviews in the client. * Added new origin in case the project is moved on render (the repository on GitHub was also slightly renamed to "idg2001_social_networking_service"). * Added/re-added helpful comments.
