+++ app/main.py
@@ -11,6 +11,7 @@ app.add_middleware(
     ## Allowed origins, frontend during development, more when deploying.
     allow_origins=[
         "http://127.0.0.1:5500",
+        "https://idg2001-social-networking-service.onrender.com",
         "https://idg2001-o1-social-networking-service.onrender.com"
     ],
     allow_credentials=True,
@@ -23,6 +24,6 @@ def read_root():
     return {"message": "Server is running"}
 
 ## Including the routers from other files.
-app.include_router(users.router)
+app.include_router(users.router, prefix="/users", tags=["Users"])
 app.include_router(posts.router, prefix="/posts", tags=["Posts"])
-app.include_router(auth_routes.router)
+app.include_router(auth_routes.router, prefix="/auth", tags=["Auth"])
