+++ app/routes/auth_routes.py
@@ -9,8 +9,9 @@ from app import crud
 from app.models.models import LoginInput, SignupInput
 from app.session import create_session, get_user_id, delete_session, session_store
 import uuid
+## Note that directories are separated by a dot (.) and not a slash (/).
 
-router = APIRouter(prefix="/auth", tags=["auth"])
+router = APIRouter()
 
 ###############
 ### Signup  ###
